Separate connection management from auth
Move VUU connection registry and session resolver code into a dedicated connection-management module while preserving the public core exports.\n\nCo-authored-by: Copilot App <223556219+Copilot@users.noreply.github.com>
