+++ vuu-ui/packages/core/docs/auth.md
@@ -1,3 +1,6 @@
 # @vuu-ui/core authentication
 
 Authentication contracts and React integration for VUU UI applications.
+
+Connection lifecycle, connection sharing, and VUU session resolution live in
+`src/connection-management`.
