+++ vuu-ui/packages/core/src/connection-management/VuuConnectionRegistry.ts
@@ -3,12 +3,12 @@ import {
   type ConnectionStatus,
   type VuuConnectionResult,
 } from "@vuu-ui/vuu-data-remote";
-import type { AuthHandler } from "./AuthHandler";
+import type { AuthHandler } from "../auth/AuthHandler";
 import {
   exchangeVuuToken,
   type VuuAuthTarget,
   type VuuSession,
-} from "./VuuTokenExchange";
+} from "../auth/VuuTokenExchange";
 import {
   IdentityTokenSessionResolver,
   type VuuSessionResolver,
